migrate user pages
