--- ui/src/app/users/ViewUserProfile.tsx
@@ -1,9 +1,0 @@
-import { Box, Heading } from "@chakra-ui/react";
-
-export default function ViewUserProfile() {
-    return (
-        <Box>
-            <Heading>View Profile: </Heading>
-        </Box>
-    )
-}
