--- ui/src/app/users/index.tsx
@@ -1,9 +1,0 @@
-import { Box, Heading } from "@chakra-ui/react";
-
-export default function ListUsers() {
-    return (
-        <Box>
-            <Heading>Users</Heading>
-        </Box>
-    )
-}
