+++ ui/src/routes/users/index.tsx
@@ -1,9 +1,12 @@
-import { createFileRoute } from '@tanstack/react-router'
+import { createFileRoute } from "@tanstack/react-router";
+import { Box, Heading } from "@chakra-ui/react";
 
-export const Route = createFileRoute('/users/')({
-  component: RouteComponent,
-})
+export const Route = createFileRoute("/users/")({ component: RouteComponent });
 
 function RouteComponent() {
-  return <div>Hello "/users/"!</div>
+  return (
+    <Box>
+      <Heading>Users</Heading>
+    </Box>
+  );
 }
