+++ ui/src/routes/users/view/$userID.tsx
@@ -1,9 +1,14 @@
-import { createFileRoute } from '@tanstack/react-router'
+import { createFileRoute } from "@tanstack/react-router";
+import { Box, Heading } from "@chakra-ui/react";
 
-export const Route = createFileRoute('/users/view/$userID')({
+export const Route = createFileRoute("/users/view/$userID")({
   component: RouteComponent,
-})
+});
 
 function RouteComponent() {
-  return <div>Hello "/users/view/$userID"!</div>
+  return (
+    <Box>
+      <Heading>View Profile: </Heading>
+    </Box>
+  );
 }
